\name{print.basad}
\alias{print.basad}
\title{Print Summary Output of Analysis}
\description{
    Print summary output from spike and slab analysis.
    Note that this is the default print method for the package.
}
\usage{
\method{print}{basab}(x, ...)
}
\arguments{
    \item{x}{An object of class \code{basad}.}
    \item{...}{Further arguments passed to or from other methods.}

}
\details{

}
\references{
Narisetty, N. N., & He, X. (2014). Bayesian variable selection with shrinking and diffusing priors. \emph{The Annals of Statistics}, 42(2), 789-817.
}
\examples{

}
\author{

Qinyan Xiang (\email{qyxiang16@gmail.com})

Naveen Narisetty ( \email{naveen@illinois.edu} )

}
\keyword{print}
